+++ CHANGELOG.md
@@ -77,3 +77,4 @@
 * Added benchmarks
 * Added back point dipole sources (direct and octree)
 * Added Kelvin force calculation (more stable than Maxwell forces currently)
+* Added two magnetized spheres example